+++ packages/component/src/ScrollToBottom/InternalContext.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const context = React.createContext({
+  _handleUpdate: () => 0,
+  _setTarget: () => 0
+});
+
+context.displayName = 'ScrollToBottomInternalContext';
+
+export default context;